+++ index.js
@@ -1,7 +1,7 @@
 // Reflects ./ts.js.
 
 module.exports = {
-  "extends": [
-    "./ts.js"
+  extends: [
+    './ts.js',
   ],
 };