+++ react-common.js
@@ -3,29 +3,29 @@
 // add further configs
 
 module.exports = {
-  "settings": {
-    "react": {
-      "version": "detect", // To remove the warning.
-    }
+  settings: {
+    react: {
+      version: 'detect', // To remove the warning.
+    },
   },
-  "plugins": [
-    "react",
+  plugins: [
+    'react',
   ],
-  "extends": [
-    "plugin:react/recommended",
-    "plugin:react-hooks/recommended",
-    "plugin:react/jsx-runtime", // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/react-in-jsx-scope.md#when-not-to-use-it
+  extends: [
+    'plugin:react/recommended',
+    'plugin:react-hooks/recommended',
+    'plugin:react/jsx-runtime', // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/react-in-jsx-scope.md#when-not-to-use-it
   ],
-  "rules": {
-    "react/prop-types": "off",
-    "react/display-name": "off",
+  rules: {
+    'react/prop-types': 'off',
+    'react/display-name': 'off',
     'react/jsx-indent': ['warn', 2, { checkAttributes: true, indentLogicalExpressions: true }],
     /** No space between prop and equal, and equal and value.
      *
      * `Wrong: <C prop = {value}/>; Right: <C prop={value} />`
      *
      * https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-equals-spacing.md */
-    "react/jsx-equals-spacing": ["warn", "never"],
+    'react/jsx-equals-spacing': ['warn', 'never'],
     /** No space inside prop value, and no space inside children brackets.
      *
      * `Wrong: <C prop={ value }/>; Right: <C prop={value} />`
@@ -33,7 +33,7 @@ module.exports = {
      * `Wrong: <C>{ value }</C>; Right: <C>{value}</C>`
      *
      * https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-curly-spacing.md */
-    "react/jsx-curly-spacing": ["warn", { "when": "never", "children": true }],
+    'react/jsx-curly-spacing': ['warn', { when: 'never', children: true }],
 
     /** No spaces after tag opening and before tag closing.
      *
@@ -42,22 +42,22 @@ module.exports = {
      * Wrong: <View />; Right: <View/>
      *
      * https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-tag-spacing.md */
-    "react/jsx-tag-spacing": ["warn", {
-      "beforeSelfClosing": "never",
-      "beforeClosing": "never"
+    'react/jsx-tag-spacing': ['warn', {
+      beforeSelfClosing: 'never',
+      beforeClosing: 'never',
     }],
 
     /** Allow to use <Component children={x}/>. Sometimes it's useful to simplify the code and to do some nice hacks,
      * like conditional wrapping component.
      * https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-children-prop.md */
-    "react/no-children-prop": "off",
+    'react/no-children-prop': 'off',
 
     /** Enable deps autofix. I've only had infinite loops once, and the autofix was off,
     * I just accept the suggestions and don't spend time analysing it. Rare to happen, easy to notice.
     *
     * https://github.com/facebook/react/issues/18235#issuecomment-898636301 */
-    "react-hooks/exhaustive-deps": ["warn", {
-      "enableDangerousAutofixThisMayCauseInfiniteLoops": true
+    'react-hooks/exhaustive-deps': ['warn', {
+      enableDangerousAutofixThisMayCauseInfiniteLoops: true,
     }],
     /** https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-no-useless-fragment.md */
     // Dropped it as it won't allow <></> to fit JSX.Element return.
@@ -69,14 +69,14 @@ module.exports = {
      * so we use `jsx-quotes` rule for props and `quotes` rule for children.
      *
      * https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-curly-brace-presence.md */
-    "react/jsx-curly-brace-presence": ["warn", { "props": "never", "children": "always"}],
+    'react/jsx-curly-brace-presence': ['warn', { props: 'never', children: 'always' }],
 
     /** Convert <A b="c"/> to <A b='c'/>. */
-    "jsx-quotes": ["warn", "prefer-single"],
+    'jsx-quotes': ['warn', 'prefer-single'],
 
     /** Prefer <> over <React.Fragment> if not using key=''.
      *
      * https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/jsx-fragments.md */
-    "react/jsx-fragments": "warn"
-  }
+    'react/jsx-fragments': 'warn',
+  },
 };